refactor: remove reliance on group.js, which make code much harder to read
