--- src/compiler/group.js
@@ -1,25 +1,0 @@
-'use strict';
-
-module.exports = {
-  def: groupdef
-};
-
-function groupdef(name, opt) {
-  opt = opt || {};
-  return {
-    _name: name || undefined,
-    type: 'group',
-    from: opt.from,
-    properties: {
-      enter: {
-        x: opt.x || undefined,
-        y: opt.y || undefined,
-        width: opt.width || {group: 'width'},
-        height: opt.height || {group: 'height'}
-      }
-    },
-    scales: opt.scales || undefined,
-    axes: opt.axes || undefined,
-    marks: opt.marks || []
-  };
-}
